Add objectstore warning

+++ leptonai/cli/objectstore.py
@@ -1,6 +1,7 @@
 """
 KV is a module that provides a way to manage kvs on a workspace.
 """
+import sys
 
 import click
 from datetime import datetime
@@ -26,7 +27,12 @@ def objectstore():
     pass
 
 
-@objectstore.command()
+@objectstore.command(
+    context_settings=dict(
+        ignore_unknown_options=True,
+        allow_extra_args=True,
+    )
+)
 @click.option("--key", "-k", help="Object key", type=str, required=True)
 @click.option("--file", "-f", help="File to write to.", type=str, default=None)
 @click.option("--public", "-p", is_flag=True, default=False, help="Is public bucket")
@@ -34,15 +40,23 @@ def objectstore():
     "--return-url",
     "-u",
     help=(
-        "Return the url of the object instead of the content. If specified, --file"
-        " takes no effect."
+            "Return the url of the object instead of the content. If specified, --file"
+            " takes no effect."
     ),
     is_flag=True,
 )
-def get(key, file, public, return_url):
+@click.pass_context
+def get(ctx, key, file, public, return_url):
     """
     Gets the object with the given key.
     """
+    # backward warning.
+    if "--bucket" in ctx.args or "-b" in ctx.args:
+        console.print(
+            "The '--bucket public' and '--bucket private' options for the 'lep photon objectstore get' command are "
+            "deprecated. Please use the '--public' option to get public objects. By default, will get private objects."
+        )
+        sys.exit(1)
     if not file:
         file = os.path.basename(key)
     client = APIClient()
@@ -63,13 +77,24 @@ def get(key, file, public, return_url):
     )
 
 
-@objectstore.command()
+@objectstore.command(context_settings=dict(
+    ignore_unknown_options=True,
+    allow_extra_args=True,
+))
 @click.option("--key", "-k", help="Object key", type=str, required=True)
 @click.option("--public", is_flag=True, default=False, help="Is public bucket")
-def cat(key, public):
+@click.pass_context
+def cat(ctx, key, public):
     """
     Gets the object with the given key and prints it to stdout.
     """
+    if "--bucket" in ctx.args or "-b" in ctx.args:
+        console.print(
+            "The '--bucket public' and '--bucket private' options for the 'lep photon objectstore cat' command are "
+            "deprecated. Please use the '--public' option to cat public objects. By default, will cat private objects."
+        )
+        sys.exit(1)
+
     client = APIClient()
 
     response = client.object_storage.get(key, public, stream=True)
@@ -78,7 +103,10 @@ def cat(key, public):
             print(chunk.decode(), end="")
 
 
-@objectstore.command()
+@objectstore.command(context_settings=dict(
+    ignore_unknown_options=True,
+    allow_extra_args=True,
+))
 @click.option(
     "--key",
     "-k",
@@ -87,10 +115,18 @@ def cat(key, public):
 )
 @click.option("--file", "-f", help="File to upload", type=str, required=True)
 @click.option("--public", is_flag=True, default=False, help="Is public bucket")
-def put(key, file, public):
+@click.pass_context
+def put(ctx, key, file, public):
     """
     Puts the object with the given key.
     """
+    if "--bucket" in ctx.args or "-b" in ctx.args:
+        console.print(
+            "The '--bucket public' and '--bucket private' options for the 'lep photon objectstore put' command are "
+            "deprecated. Please use the '--public' option to put public objects. By default, private objects "
+            "will be put."
+        )
+        sys.exit(1)
     if not key:
         key = file
     if os.path.getsize(file) > _max_upload_file_size_limit:
@@ -108,27 +144,50 @@ def put(key, file, public):
         console.print(f"Successfully uploaded object [green]{key}[/].")
 
 
-@objectstore.command()
+@objectstore.command(context_settings=dict(
+    ignore_unknown_options=True,
+    allow_extra_args=True,
+)
+)
 @click.option("--key", "-k", help="Object key", type=str, required=True)
 @click.option("--public", is_flag=True, default=False, help="Is public bucket")
-def delete(key, public):
+@click.pass_context
+def delete(ctx, key, public):
     """
     Deletes the object with the given key.
     """
-
+    if "--bucket" in ctx.args or "-b" in ctx.args:
+        console.print(
+            "The '--bucket public' and '--bucket private' options for the 'lep photon objectstore delete' command are "
+            "deprecated. Please use the '--public' option to delete public objects. By default, private objects "
+            "will be deleted."
+        )
+        sys.exit(1)
     client = APIClient()
     client.object_storage.delete(key, public)
     console.print(f"Successfully deleted object [green]{key}[/].")
 
 
-@objectstore.command(name="list")
+@objectstore.command(name="list",
+                     context_settings=dict(
+                         ignore_unknown_options=True,
+                         allow_extra_args=True,
+                     )
+                     )
 @click.option("--prefix", "-p", help="Prefix to filter objects", type=str, default=None)
 @click.option("--public", is_flag=True, default=False, help="Is public bucket")
-def list_command(prefix, public):
+@click.pass_context
+def list_command(ctx, prefix, public):
     """
     Lists all objects in the current workspace.
     """
-
+    if "--bucket" in ctx.args or "-b" in ctx.args:
+        console.print(
+            "The '--bucket public' and '--bucket private' options for the 'lep photon objectstore list' command are "
+            "deprecated. Please use the '--public' option to list public objects. By default, private objects "
+            "will be listed."
+        )
+        sys.exit(1)
     client = APIClient()
 
     storage_metadatas = client.object_storage.list(prefix, public)
